Move QueryDataEncodingModule to spooling package
