+++ core/trino-main/src/main/java/io/trino/server/protocol/spooling/QueryDataEncodingModule.java
@@ -11,14 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.server.protocol.spooling.encoding;
+package io.trino.server.protocol.spooling;
 
 import com.google.inject.Binder;
 import com.google.inject.Scopes;
 import com.google.inject.multibindings.Multibinder;
 import io.airlift.configuration.AbstractConfigurationAwareModule;
-import io.trino.server.protocol.spooling.QueryDataEncoder;
-import io.trino.server.protocol.spooling.QueryDataEncoders;
+import io.trino.server.protocol.spooling.encoding.JsonQueryDataEncoder;
 
 import static com.google.inject.multibindings.Multibinder.newSetBinder;
